+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## Unreleased
+
+### Changed
+
+ - Package `maska` version updated to `3.0.3`
+ 
+### Fixed
+
+- `click-outside` directive syntax
+
 ## v3.0.0
 
 ### Changed